[HOTSPOT-181] feat: 구성원 데이터 한도, 즉시 차단 여부 예외 처리 로직 추가

+++ src/main/java/hotspot/user/family/controller/response/UpdateDataLimitResponse.java
@@ -13,7 +13,7 @@ import lombok.Builder;
 public record UpdateDataLimitResponse(
         Long familyId,
         Long subId,
-        long dataLimit,
+        double dataLimit,
         boolean isLocked
 ) {
 }
